+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
+## [0.0.3]
+- Fixed - Each project now has its own files list json file.
+
 ## [0.0.2]
-- Added `Version Inc: Edit Settings for Update Files`
+- Added - `Version Inc: Edit Settings for Update Files`
 
 ## [0.0.1]